requirements.txt
sk-video==1.1.10
opencv_python==4.4.0.42
av==8.0.2
tqdm==4.48.2
dlib==19.21.0
streamlit==0.62.0
torchvision==0.7.0
tensorflow_gpu==1.15.0
scipy==1.5.2
numpy==1.19.1
torch==1.4.0
Pillow==7.2.0
